move stuff into output/
