+++ src/lib.rs
@@ -1,10 +1,11 @@
-use crate::fmt_md::MdOptions;
-use crate::fmt_md_inlines::MdInlinesWriterOptions;
-use crate::fmt_plain::PlainOutputOpts;
 use crate::md_elem::{InvalidMd, MdDoc, MdElemRef, MdSerde, ParseOptions};
 use crate::query::ParseError;
 use crate::select::{SelectorAdapter, SelectorChain};
 use cli::{Cli, OutputFormat};
+use output::fmt_md::MdOptions;
+use output::fmt_md_inlines::MdInlinesWriterOptions;
+use output::fmt_plain::PlainOutputOpts;
+use output::{fmt_md, fmt_plain};
 use pest::error::ErrorVariant;
 use pest::Span;
 use std::borrow::Cow;
@@ -16,15 +17,9 @@ use util::output::Output;
 use util::output::{OutputOpts, Stream};
 
 pub mod cli;
-mod fmt_md;
-mod fmt_md_inlines;
-mod fmt_plain;
-mod fmt_plain_writer;
-mod fmt_str;
-mod footnote_transform;
-mod link_transform;
 mod matcher;
 pub mod md_elem;
+mod output;
 mod query;
 mod select;
 mod util;
