+++ src/output/fmt_plain.rs
@@ -1,7 +1,7 @@
-use crate::fmt_plain_writer::NewlineCollapser;
 use crate::md_elem::elem::*;
 use crate::md_elem::elem_ref::*;
 use crate::md_elem::*;
+use crate::output::fmt_plain_writer::NewlineCollapser;
 use std::io::{Error, LineWriter, Write};
 
 #[derive(Copy, Clone, Debug, PartialEq, Eq)]
