+++ src/output/mod.rs
@@ -1,0 +1,7 @@
+pub mod fmt_md;
+pub mod fmt_md_inlines;
+pub mod fmt_plain;
+mod fmt_plain_writer;
+pub mod fmt_str;
+mod footnote_transform;
+pub mod link_transform;
